+++ oldlx/Trees/ShrubModel.java
@@ -254,9 +254,6 @@ class ShrubConfig  {
   float x;
   float z;
   float ry;
-  //    int[] canopyMajorLengths;
-  //    int[] clusterBaseHeights;
-  //    String ipAddress;
 }
 
 class Shrub extends LXModel {
