+++ src/main/java/com/charlesgadeken/entwined/EntwinedGui.java
@@ -27,6 +27,7 @@ public class EntwinedGui extends PApplet implements LXPlugin {
 
     @Override
     public void setup() {
+
         LXStudio.Flags flags = new LXStudio.Flags(this);
         flags.resizable = true;
         flags.useGLPointCloud = false;
