+++ src/main/java/com/charlesgadeken/entwined/model/Cube.java
@@ -1,0 +1,24 @@
+package com.charlesgadeken.entwined.model;
+
+import com.charlesgadeken.entwined.model.config.CubeConfig;
+import toxi.geom.Vec3D;
+
+public class Cube extends BaseCube {
+    public static final int[] PIXELS_PER_CUBE = {
+        6, 6, 6, 12, 12
+    }; // Tiny cubes actually have less, but for Entwined we want to
+    // tell the NDB that everything is 6
+    public static final float[] CUBE_SIZES = {4f, 7.5f, 11.25f, 15f, 16.5f};
+    /** Size of this cube, one of SMALL/MEDIUM/LARGE/GIANT */
+    public final float size;
+
+    public final int pixels;
+    public CubeConfig config = null;
+
+    Cube(Vec3D globalPosition, Vec3D treePosition, CubeConfig config) {
+        super(globalPosition, treePosition);
+        this.size = CUBE_SIZES[config.cubeSizeIndex];
+        this.pixels = PIXELS_PER_CUBE[config.cubeSizeIndex];
+        this.config = config;
+    }
+}
