+++ src/main/java/com/charlesgadeken/entwined/model/Geometry.java
@@ -1,0 +1,6 @@
+package com.charlesgadeken.entwined.model;
+
+public class Geometry {
+    static final int INCHES = 1;
+    static final int FEET = 12 * INCHES;
+}
