+++ src/main/java/com/charlesgadeken/entwined/model/Layer.java
@@ -1,0 +1,13 @@
+package com.charlesgadeken.entwined.model;
+
+import heronarts.lx.LX;
+import heronarts.lx.LXLayer;
+
+public abstract class Layer extends LXLayer {
+    protected final Model model;
+
+    Layer(LX lx) {
+        super(lx);
+        model = (Model) lx.model;
+    }
+}
