+++ src/main/java/com/charlesgadeken/entwined/model/ModelTransformTask.java
@@ -1,0 +1,16 @@
+package com.charlesgadeken.entwined.model;
+
+import heronarts.lx.LXLoopTask;
+
+public class ModelTransformTask implements LXLoopTask {
+    protected final Model model;
+
+    ModelTransformTask(Model model) {
+        this.model = model;
+    }
+
+    @Override
+    public void loop(double deltaMs) {
+        model.runTransforms();
+    }
+}
