+++ src/main/java/com/charlesgadeken/entwined/model/TreeConfig.java
@@ -1,0 +1,9 @@
+package com.charlesgadeken.entwined.model;
+
+public class TreeConfig {
+    public float x;
+    public float z;
+    public float ry;
+    public int[] canopyMajorLengths;
+    public int[] layerBaseHeights;
+}
