+++ src/main/java/com/charlesgadeken/entwined/model/config/CubeConfig.java
@@ -1,20 +1,20 @@
 package com.charlesgadeken.entwined.model.config;
 
 public class CubeConfig {
-    int sculptureIndex;
-    int cubeSizeIndex;
-    int outputIndex;
-    String ipAddress;
+    public int sculptureIndex;
+    public int cubeSizeIndex;
+    public int outputIndex;
+    public String ipAddress;
 
     // For Tree
-    int treeIndex;
-    int layerIndex;
-    int branchIndex;
-    int mountPointIndex;
-    boolean isActive;
+    public int treeIndex;
+    public int layerIndex;
+    public int branchIndex;
+    public int mountPointIndex;
+    public boolean isActive;
 
     // For Shrub
-    int shrubIndex;
-    int clusterIndex;
-    int rodIndex;
+    public int shrubIndex;
+    public int clusterIndex;
+    public int rodIndex;
 }
