+++ src/main/java/com/charlesgadeken/entwined/model/config/ShrubCubeConfig.java
@@ -1,12 +1,12 @@
 package com.charlesgadeken.entwined.model.config;
 
 public class ShrubCubeConfig {
-    int shrubIndex; // each shrubIndex maps to an ipAddress, consider pushing ipAddress up to
+    // each shrubIndex maps to an ipAddress, consider pushing ipAddress up to
+    public int shrubIndex;
     // ShrubConfig
-    int clusterIndex;
-    int rodIndex;
-    //    int mountPointIndex;
-    int shrubOutputIndex;
-    int cubeSizeIndex;
-    String shrubIpAddress;
+    public int clusterIndex;
+    public int rodIndex;
+    public int shrubOutputIndex;
+    public int cubeSizeIndex;
+    public String shrubIpAddress;
 }
